cdk/requirements.txt
# AWS CDK Core
aws-cdk-lib>=2.232.0
constructs>=10.0.0

# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# Utilities
pyyaml>=6.0